aws_infra/multinode_batch_cdk/requirements.txt
boto3
aws-cdk-lib==2.40.0
constructs>=10.0.0,<11.0.0
